pyproject.toml
[project]
name = "classgen"
version = "4.0.0"
description = "AI lesson activity generator for African secondary school teachers"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135",
    "uvicorn>=0.41",
    "openai>=2.24",
    "supabase>=2.28",
    "twilio>=9.10",
    "fpdf2>=2.8",
    "python-dotenv>=1.2",
    "python-multipart>=0.0.22",
    "jinja2>=3.1.6",
    "pywebpush>=2.3.0",
    "babel>=2.16",
    "redis>=5.0",
    "psycopg[binary]>=3.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=7.0",
    "httpx>=0.28",
    "ruff>=0.15.6",
    "pyright>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/classgen"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
# System prompts are long text constants — line length is irrelevant
"src/classgen/content/prompts.py" = ["E501"]
# Tests have long mock paths and inline fixtures — readability over line length
"tests/**" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
